Add Espresso types and define justification required for L2 batches
Define the HotShot types that the OP derivation pipeline will need
to work with -- block headers and related proof types -- and define
the extra information that must be included in a batch to show
that that batch has been correctly constructed from a range of
HotShot blocks.

The L2BatchJustification type is added to the OP chain in two
places:
1. It is added to BatchInfo, so that it can be serialized and sent
   to the batch inbox on L1. This preserves the invariant that the
   entire L2 chain can be derived from information present on L1.
2. It is added to the calldata of the system-generated L1 info
   deposit transaction. This is really just a convenient place to
   put it so that it gets serialized into the L2 block without
   changing the L2 block type (which is currently identical to the
   L1 block type). Doing this allows us to reconstruct batches
   _with justification_ from blocks, which is important because
   the batches submitted by the batcher are derived from blocks
   produced by the execution engine.
By adding the justification to both the batch and the block, we
ensure that it will be correctly serialized and sent to L1 by the
batcher, then deserialized and recovered from L1 by the derivation
pipeline, without requiring further changes to either (except the
business logic of populating the justification in the batcher and
validating the batch constraints in the batch queue).

This change attempts to preserve support for the old BatchV1 type
without this justification, so that
* We can still run the node without Espresso
* The node will be able to replay a chain which started before
  Espresso and then added support for Espresso some time later

I had to change many, many tests, but on the bright side one of
those tests caught an actual bug -- the `nil` edge case in
`L1BlockInfo.MarshalBinary()`.

The next step is to implement the basic operations that we will
need on the HotShot types. Since NMT proof validation will be
mocked for a while, the main operation we need is the ability to
compute a HotShot block commitment from a `Header`, so that we can
authenticate the headers in the justification against certified
block commitments from the sequencer contract.

This change closes #15.