Add recipe to make bindings in docker
This should generate the same bindings like the CI because it's running
in container made from the same image as long as the image isn't
updated between the runs.

For the build inside docker we pass `--force` to `forge build` to avoid
hitting the cache. Unfortunately using a different foundry version does
not seem to invalidate the cache.