Add Espresso flag to SystemConfig
The flag determines whether an OP-chain is using the Espresso
sequencer, which changes the way the OP node works in sequencer
mode and changes the derivation pipeline. Adding this to SystemConfig
allows the sequencer and batch queue to read this information when
they need to. It also allows a chain to dynamically opt in or out
of using the Espresso Sequencer by calling a method on the L1
SystemConfig contract. Maybe. I don't plan to test this for Cortado

More importantly for our purposes, it gives us a way to switch
between sequencing modes at config time, so that in the default
config, the sequencer and derivation pipeline will continue
working the way they always have. This means the all the existing
tests will continue to work, and we won't have to constantly update
them during development. We will only have to add new tests to test
the behavior of the node when Espresso mode is switched on.

By keeping the existing code around mostly unmodified, and adding
new code to support the mode where Espresso is enabled, we also
make it easier to keep this fork in sync with the upstream code
base.