Set up e2e tests to use Espresso mode
Running the tests with both legacy mode and Espresso mode in the
same process is problematic, because the e2e tests use global
variables initialized based on command line arguments to get the
system config and contract deployments, both of which must change
depending on the mode.

Instead, the test code does not change at all, and we invoke
`go test` twice, with different command line arguments, to run the
tests using an Espresso deployment and a legacy deployment.

The devnet Python script has been parameterized based on a directory
for output files, and there is a new make target to generate
contract deployments and L1 genesis state with the `Espresso` flag
enabled.

CircleCI jobs have been added to run the e2e tests in Espresso mode.
These jobs will not pass until we finish implementing the sequencer
(https://github.com/EspressoSystems/op-espresso-integration/issues/30)
and then we will probably have to selectively enable/disable certain
tests that we want to run with Espresso.