Implement espresso.Client
With this change, I can run TestSystemE2E and see Espresso batches
successfully get sequenced, although the test later crashes in
the derivation pipeline on the unimplemented `CheckBatchEspresso`.
Also, there are no transactions in the batches, because I need to
change the test to use the proxy RPC as its L2 client, so that
transaction submissions go to Espresso.

Closes #30