Make Espresso e2e setup more reliable
We start up the Espresso system and wait for it to show progress
before starting the L2 chains. Otherwise, the L2 chain is initialized
with the L1 origin from before we start up Espresso, and by the
time Espresso is up 10ish seconds later, the L2's view of L1 is
way behind Espresso, which causes tests to time out while the L2
is catching up.

The new setup process more closely matches the deployment order for
a real system:
- The L1 exists first
- Espresso is deployed
- OP chains are deployed in Espresso mode (or opt into Espresso mode)
  _after_ the Espresso ledger already exists