Improve selection of L1 origin in Espresso mode
We now apply corrections to the L1 origin _before_ deciding
whether the batch must be empty, which should lead to fewer
spurious empty batches.

We also use the L1 head of the _last_ Espresso block instead of the
first one, which gives us a more recent L1 origin. Since we must
always wait for the last block of the window to choose the L1
origin, the justification format and validation has fewer edge
cases. The justification will always have start and end bookends,
and the validation flow is simply:
* validate all headers
* check bookends
* check L1 origin
* check transactions
The only edge cases are in checking the bookends, where Prev may be
nil in the genesis case.

Closes #52
Closes #53