Add commands to deploy a new L2 onto an existing L1 and sequencer
The second L2 works differently than the first one, which is built into
the L1 genesis block. Deploying the second L2 onto an existing L1 and
sequencer is simpler than trying to build two different L2 deployments
into the L1 genesis, and it showcases a different, and perhaps more
realistic, workflow: launching a new L2 while a running Espresso
Sequencer already exists. In particular, we demonstrate that no changes
to the sequencer are required.