Increase L2 output oracle submission interval for Espreso devnets
Parker from Caldera discovered that the interval was too low, so that L2 blocks
were being generated faster than the OP proposer could keep up, resulting in long
latencies for withdrawals to be proven that increased as the devnet went on. With
this change, the time to run `make devnet-test-espresso` (which includes proving a
withdrawal) decreases from 12m to 3:25 on my machine.